Rename _app.js since we removed next.js
